move post_util to where its used
